@@ -1,89 +1,84 @@
 ﻿using ASPNETCoreMVCSecurity.Models;
+
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Data.SqlClient;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
+
 using System.Diagnostics;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace ASPNETCoreMVCSecurity.Controllers
+namespace ASPNETCoreMVCSecurity.Controllers;
+
+public class HomeController : Controller
 {
-    public class HomeController : Controller
+    private readonly ILogger<HomeController> _logger;
+    private readonly IConfiguration _settings;
+
+    public HomeController(ILogger<HomeController> logger, IConfiguration configuration)
     {
-        private readonly ILogger<HomeController> _logger;
-        private readonly IConfiguration _settings;
+        _logger = logger;
+        _settings = configuration;
+    }
 
-        public HomeController(ILogger<HomeController> logger, IConfiguration configuration)
+    public IActionResult Index()
+    {
+        return View();
+    }
+
+    public IActionResult Privacy()
+    {
+        return View();
+    }
+
+    public string Echo(string x) => x;
+
+    public IActionResult EchoUnencoded(string x) => Content(x, "text/html");
+
+    public IActionResult EchoWithView(string x)
+    {
+        ViewBag.SampleData = x;
+        return View();
+    }
+
+    public IActionResult SqlSample(string id)
+    {
+        string connectionString = _settings.GetConnectionString("InjectionConnection");
+        SqlConnection sqlConnection = new(connectionString);
+        SqlCommand command = sqlConnection.CreateCommand();
+
+        // don't do this - string concatenation for SQL commands!
+        command.CommandText = "SELECT * FROM Customers WHERE City = " + id;
+        sqlConnection.Open();
+        using SqlDataReader reader = command.ExecuteReader(System.Data.CommandBehavior.CloseConnection);
+
+        StringBuilder sb = new();
+        while (reader.Read())
         {
-            _logger = logger;
-            _settings = configuration;
-        }
-
-        public IActionResult Index()
-        {
-            return View();
-        }
-
-        public IActionResult Privacy()
-        {
-            return View();
-        }
-
-        public string Echo(string x) => x;
-
-        public IActionResult EchoUnencoded(string x) => Content(x, "text/html");
-
-        public IActionResult EchoWithView(string x)
-        {
-            ViewBag.SampleData = x;
-            return View();
-        }
-
-        public IActionResult SqlSample(string id)
-        {
-            string connectionString = _settings.GetConnectionString("InjectionConnection");
-            SqlConnection sqlConnection = new(connectionString);
-            SqlCommand command = sqlConnection.CreateCommand();
-
-            // don't do this - string concatenation for SQL commands!
-            command.CommandText = "SELECT * FROM Customers WHERE City = " + id;
-            sqlConnection.Open();
-            using SqlDataReader reader = command.ExecuteReader(System.Data.CommandBehavior.CloseConnection);
-
-            StringBuilder sb = new();
-            while (reader.Read())
+            for (int i = 0; i < reader.FieldCount; i++)
             {
-                for (int i = 0; i < reader.FieldCount; i++)
-                {
-                    sb.Append(reader[i]);
-                }
-                sb.AppendLine();
+                sb.Append(reader[i]);
             }
-            ViewBag.Data = sb.ToString();
-
-            return View();
+            sb.AppendLine();
         }
+        ViewBag.Data = sb.ToString();
 
-        public IActionResult EditBook() => View();
+        return View();
+    }
 
-        [HttpPost]
-        public IActionResult EditBook(Book book) => View("EditBookResult", book);
+    public IActionResult EditBook() => View();
 
-        public IActionResult EditBookSecure() => View();
+    [HttpPost]
+    public IActionResult EditBook(Book book) => View("EditBookResult", book);
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult EditBookSecure(Book book) => View("EditBookResult", book);
+    public IActionResult EditBookSecure() => View();
+
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public IActionResult EditBookSecure(Book book) => View("EditBookResult", book);
 
 
-        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error()
-        {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-        }
+    [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+    public IActionResult Error()
+    {
+        return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
     }
 }
